Fix: Bring back old locale switcher
