+++ apps/web/src/app/[locale]/settings/language/page.tsx
@@ -1,4 +1,4 @@
-import LocaleSwitcher from "@/components/Shared/LocaleSwitcher";
+import LanguageSetting from "@/components/Shared/LanguageSetting";
 import PageBottom from "@/components/Shared/PageBottom";
 import PageTitle from "@/components/Shared/PageTitle";
 import { loadMessages } from "@/lib/internationalisation/i18n";
@@ -20,7 +20,7 @@ export default async function Language({
         titleOnly={true}
         backHref={`/${locale}/settings`}
       />
-      <LocaleSwitcher locale={locale} />
+      <LanguageSetting locale={locale} />
       <PageBottom />
     </main>
   );
